Typelo/Backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
redis>=5.0.1
firebase-admin>=6.5.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
motor>=3.3.2
pymongo>=4.6.1
bcrypt>=4.0.0